+++ services/celery_services/pay_app.py
@@ -1,0 +1,12 @@
+"""
+pip install celery==5.2.3
+"""
+
+from celery import Celery
+
+from .config import CeleryConfig
+
+
+app = Celery('tasks')
+
+app.config_from_object(CeleryConfig)